+++ BALSAMIC/commands/init/base.py
@@ -1,26 +1,26 @@
-import os
-import sys
-import re
 import logging
+import re
 import subprocess
+import sys
 from pathlib import Path
 
 import click
 import graphviz
 import snakemake
 
+from BALSAMIC import __version__ as balsamic_version
 from BALSAMIC.constants.analysis import (
     BIOINFO_TOOL_ENV,
 )
 from BALSAMIC.constants.cache import DOCKER_URL, DockerContainers
+from BALSAMIC.constants.cluster import ClusterConfigType
 from BALSAMIC.utils.cli import (
     CaptureStdout,
     get_snakefile,
     SnakeMake,
-    get_config,
     get_schedulerpy,
+    get_config_path,
 )
-from BALSAMIC import __version__ as balsamic_version
 from BALSAMIC.utils.io import write_json
 
 LOG = logging.getLogger(__name__)
@@ -105,7 +105,7 @@ LOG = logging.getLogger(__name__)
 @click.option(
     "--cluster-config",
     show_default=True,
-    default=get_config("reference_cluster"),
+    default=get_config_path(ClusterConfigType.CACHE),
     type=click.Path(),
     help="cluster config json file. (eg- SLURM, QSUB)",
 )